Updated to CCLv3

+++ ClLike/cl_like/tests/test_bias_models.py
@@ -15,7 +15,8 @@ OUTDIR = 'dum'
 
 # Cleaning the tmp dir after running the tests (even if they fail)
 def teardown_module():
-    shutil.rmtree(OUTDIR)
+    if os.path.isdir(OUTDIR):
+        shutil.rmtree(OUTDIR)
 
 
 def get_info(bias, A_sE9=True):
@@ -129,13 +130,16 @@ def get_info(bias, A_sE9=True):
             "debug": False}
 
     if not A_sE9:
+
         # info["params"]["sigma8"] = 0.78220521  # From Bacco
-        info["params"]["sigma8"] = 0.7824601264149301  # From CCL
+        # info["params"]["sigma8"] = 0.7824601264149301  # From CCLv2
+        info["params"]["sigma8"] = 0.7825789828130193  # From CCLv3
         del info["params"]["A_sE9"]
 
     return info
 
-@pytest.mark.parametrize('bias', ['Linear', 'EulerianPT', 'LagrangianPT',
+# LagrangianPT commented out because velocileptors dependency is broken
+@pytest.mark.parametrize('bias', ['Linear', 'EulerianPT', # 'LagrangianPT',
                                   'BaccoPT'])
 def test_dum(bias):
     info = get_info(bias)
@@ -214,7 +218,8 @@ def test_timing():
         # Before restructuring, the average evaluation time was ~0.54s in my laptop
         # After the restructuration, it went to 0.56s.
         # With the new data, it takes longer because there are more data
-        assert time < 2.4
+        # For CCLv3 it went a bit up too (from 2.4 to 3.4)
+        assert time < 4
 
 
 def test_null_negative_eigvals_in_icov():
@@ -383,7 +388,7 @@ def test_sigma8_to_As(case):
        "h": 0.67,
        "n_s": 0.96,
        "m_nu": 0.15,
-       "sigma8": 0.7824601264149301,
+       "sigma8": 0.7825789828130193,
        "T_CMB": 2.7255
     }
 
@@ -397,9 +402,9 @@ def test_sigma8_to_As(case):
         assert np.abs(As*1E9 / 2.1265 -1) < 1e-5
         assert np.fabs(loglikes[0]) < 3e-3
     else:
-        # The O(1e-3) difference between As makes chi2~0.37 for baccoemu
+        # The O(1e-3) difference between As makes chi2~0.77 for baccoemu
         assert np.abs(As*1E9 / 2.1265 -1) < 1e-3
-        assert np.fabs(loglikes[0]) < 0.4
+        assert np.fabs(loglikes[0]) < 0.8
 
 
 @pytest.mark.parametrize('case', ['baccoemu', 'CCL'])
